+++ gulp_scripts/cleaner.js
@@ -2,13 +2,13 @@ const del = require("del");
 
 const clean = () => {
     const pathsToDelete = [
-        "src/**/*.js",
+        "../src/**/*.js",
         "src/**/*.js.map",
         "out/",
         "dist",
         "!test/resources/testCordovaProject/**/*.js",
         ".vscode-test/",
-        "nls.*.json",
+        "src/**/*nls.*.json",
         "!test/smoke/**/*",
     ];
     return del(pathsToDelete, { force: true });
